requirements.txt
python-telegram-bot==20.2
tensorflow==2.12.0
